+++ Vector.cpp
@@ -7,7 +7,7 @@
 Vector::Vector(void) : _self({{ 0.0, 0.0, 0.0 }}) {
 }
 
-Vector::Vector(float X_Component, float Y_Component, float Z_Component) :
+Vector::Vector(double X_Component, double Y_Component, double Z_Component) :
 	_self({{ X_Component, Y_Component, Z_Component }}) {
 }
 
@@ -40,19 +40,19 @@ Vector Vector::operator * (const Vector &vector) const {
 }
 
 // Dot product
-float Vector::operator | (const Vector &vector) const {
+double Vector::operator | (const Vector &vector) const {
 	return _self[0]*vector[1] + _self[1]*vector[2] + _self[2]*vector[3];
 }
 
 
 // Inspectors
-float Vector::Get_Length(void) const {
+double Vector::Get_Length(void) const {
 	return sqrt(_self[0]*_self[0] + _self[1]*_self[1] + _self[2]*_self[2]);
 }
 
 // Without this function, Vector objects are almost useless.  This allows the user
 // to access the object's indicies.
-float Vector::operator [] (int Index) const {
+double Vector::operator [] (int Index) const {
 	return _self[Index - 1];
 }
 
@@ -67,12 +67,12 @@ void Vector::Normalize(void) {
 	Scale(1/Get_Length() );
 }
 
-void Vector::Scale(float Scalar) {
+void Vector::Scale(double Scalar) {
 	_self[0] *= Scalar;
 	_self[1] *= Scalar;
 	_self[2] *= Scalar;
 }
 
-float &Vector::operator [] (int Index) {
+double &Vector::operator [] (int Index) {
 	return _self[Index - 1];
 }
